refactor: Move the `isSingletonPresentInStack` function into a more sensible location
This function is currently in `utils/test/singleton.ts` however it doesn't relate to "test".
Move it to `utils/singleton.ts`.
