+++ src/constructs/ec2/security-groups/wazuh.ts
@@ -1,6 +1,6 @@
 import { Peer } from "@aws-cdk/aws-ec2";
 import type { IVpc } from "@aws-cdk/aws-ec2";
-import { isSingletonPresentInStack } from "../../../utils/test";
+import { isSingletonPresentInStack } from "../../../utils/singleton";
 import type { GuStack } from "../../core";
 import { GuMigratingResource } from "../../core/migrating";
 import { GuBaseSecurityGroup } from "./base";
