+++ src/constructs/iam/policies/anghammarad.ts
@@ -1,4 +1,4 @@
-import { isSingletonPresentInStack } from "../../../utils/test";
+import { isSingletonPresentInStack } from "../../../utils/singleton";
 import type { GuStack } from "../../core";
 import { GuAnghammaradTopicParameter } from "../../core";
 import { GuAllowPolicy } from "./base-policy";
