+++ src/utils/test/index.ts
@@ -1,4 +1,3 @@
 export * from "./simple-gu-stack";
 export * from "./synthed-stack";
 export * from "./attach-policy-to-test-role";
-export * from "./singleton";
